backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-mail>=0.10.0",
    "gunicorn>=23.0.0",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
]
